# openapi/agent-tools.yaml
openapi: 3.1.0
info:
  title: X News Toolbox Agent Tools
  version: 1.0.0
  description: Tools for a Mind to configure and operate one deployed X News Toolbox instance. Secrets are never accepted as tool arguments.
servers:
  - url: https://YOUR_DEPLOYMENT_HOST
security:
  - bearerAuth: []
paths:
...
  /api/agent/configure:
    post:
      operationId: configureAgentPreferences
      summary: Save non-secret creator preferences and the autonomous schedule.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentConfiguration"
      responses:
        "200": { description: Configuration accepted }
        "400": { description: Invalid input or attempted extra fields }
        "409": { description: The user must connect a Mind first, or the profile changed concurrently }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    AgentConfiguration:
      type: object
      additionalProperties: false
      required: [requestId, profile, automation]
      properties:
        requestId:
          type: string
          minLength: 8
          maxLength: 80
          description: Stable idempotency key. Reuse it when retrying the same configuration.
        profile:
          type: object
          additionalProperties: false
          required: [positioning, audience, voice, boundaries]
          properties:
            positioning: { type: string, minLength: 4, maxLength: 240 }
            audience: { type: string, minLength: 4, maxLength: 240 }
            voice: { type: string, minLength: 2, maxLength: 160 }
            boundaries: { type: string, minLength: 2, maxLength: 400 }
        automation:
          type: object
          additionalProperties: false
          required: [enabled, platform, outputCount, dailyTime]
          properties:
            enabled: { type: boolean }
            platform: { type: string, enum: [x, xiaohongshu] }
            outputCount: { type: integer, minimum: 1, maximum: 5 }
            focus: { type: string, maxLength: 240 }
            dailyTime:
              type: string
              pattern: "^([01]\\d|2[0-3]):[0-5]\\d$"